+++ app/main.py
@@ -50,11 +50,11 @@ from app.api import (
     internal_utils,
     mailbox,
     media,
+    organization,
+    patient_chat,
     patient_profile_repositories,
     patient_profiles,
-    patient_chat,
     response_templates,
-    organization,
 )
 from app.data.operations import DatabaseOperations
 from app.migrations.es_from_mongodb import move_data_from_mongo_to_es