+++ devops/terraform/main.tf
@@ -1,18 +1,41 @@
+data "azurerm_key_vault" "keyvault_devops" {
+  name                = var.devops_keyvault_name
+  resource_group_name = var.devops_resource_group_name
+}
+
+data "azurerm_key_vault_secrets" "keyvault_secrets" {
+  key_vault_id = data.azurerm_key_vault.keyvault_devops.id
+}
+
+data "azurerm_key_vault_secret" "keyvault_secrets" {
+  for_each     = toset(data.azurerm_key_vault_secrets.keyvault_secrets.names)
+  name         = each.key
+  key_vault_id = data.azurerm_key_vault.keyvault_devops.id
+}
+
+data "azurerm_key_vault_certificates" "keyvault_certs" {
+  key_vault_id = data.azurerm_key_vault.keyvault_devops.id
+}
+
+data "azurerm_key_vault_certificate" "keyvault_certs" {
+  for_each     = toset(data.azurerm_key_vault_certificates.keyvault_certs.names)
+  name         = each.key
+  key_vault_id = data.azurerm_key_vault.keyvault_devops.id
+}
+
 module "tls_certificates" {
   source                   = "./modules/tls_certificate"
-  godaddy_api_key          = var.godaddy_key
-  godaddy_api_secret       = var.godaddy_secret
-  google_domains_token     = var.google_domains_token
-  ssl_certificate_password = var.ssl_certificate_password
+  godaddy_api_key          = data.azurerm_key_vault_secret.keyvault_secrets["godaddy-key"].value
+  godaddy_api_secret       = data.azurerm_key_vault_secret.keyvault_secrets["godaddy-secret"].value
+  ssl_certificate_password = data.azurerm_key_vault_secret.keyvault_secrets["ssl-certificate-password"].value
   host_name                = var.host_name
 }
 
 module "tls_certificates_2" {
   source                   = "./modules/tls_certificate"
-  godaddy_api_key          = var.godaddy_key
-  godaddy_api_secret       = var.godaddy_secret
-  google_domains_token     = var.google_domains_token
-  ssl_certificate_password = var.ssl_certificate_password
+  godaddy_api_key          = data.azurerm_key_vault_secret.keyvault_secrets["godaddy-key"].value
+  godaddy_api_secret       = data.azurerm_key_vault_secret.keyvault_secrets["godaddy-secret"].value
+  ssl_certificate_password = data.azurerm_key_vault_secret.keyvault_secrets["ssl-certificate-password"].value
   host_name                = var.host_name_2
 }
 
@@ -39,8 +62,8 @@ module "keyvault" {
   source              = "./modules/keyvault"
   keyvault_name       = "keyvault"
   resource_group_name = module.resource_group.resource_group_name
-  azure_tenant_id     = var.azure_tenant_id
-  azure_object_id     = var.azure_object_id
+  azure_tenant_id     = data.azurerm_key_vault_secret.keyvault_secrets["azure-tenant-id"].value
+  azure_object_id     = data.azurerm_key_vault_secret.keyvault_secrets["azure-object-id"].value
   subnet_id           = module.virtual_network.container_apps_subnet_id
 }
 
@@ -60,7 +83,7 @@ module "application_gateway" {
   ui_address               = "frontend.${module.container_apps_env.container_app_environment_default_domain}"
   gateway_public_ip_id     = module.public_ip.public_ip_id
   react_app_address        = "ui.${module.container_apps_env.container_app_environment_default_domain}"
-  ssl_certificate_password = var.ssl_certificate_password
+  ssl_certificate_password = data.azurerm_key_vault_secret.keyvault_secrets["ssl-certificate-password"].value
   ssl_certificate_pfx      = module.tls_certificates.certificate_pfx
   ssl_certificate_pfx_2    = module.tls_certificates_2.certificate_pfx
   host_name                = var.host_name
@@ -81,29 +104,30 @@ module "container_app_backend" {
   container_app_env_id              = module.container_apps_env.container_app_environment_id
   docker_image                      = format("%s/%s", var.container_registry_server, var.backend_container_image_tag)
   container_registry_server         = var.container_registry_server
-  container_registry_username       = var.container_registry_username
-  container_registry_password       = var.container_registry_password
-  azure_client_id                   = var.azure_client_id
-  azure_client_secret               = var.azure_client_secret
-  azure_tenant_id                   = var.azure_tenant_id
+  container_registry_username       = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-username"].value
+  container_registry_password       = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-password"].value
+  azure_client_id                   = data.azurerm_key_vault_secret.keyvault_secrets["azure-client-id"].value
+  azure_client_secret               = data.azurerm_key_vault_secret.keyvault_secrets["azure-client-secret"].value
+  azure_tenant_id                   = data.azurerm_key_vault_secret.keyvault_secrets["azure-tenant-id"].value
   keyvault_url                      = module.keyvault.keyvault_url
-  jwt_secret                        = var.jwt_secret
-  mongo_connection_url              = var.mongo_connection_url
-  outlook_client_id                 = var.outlook_client_id
-  outlook_client_secret             = var.outlook_client_secret
-  outlook_tenant_id                 = var.outlook_tenant_id
-  outlook_redirect_uri              = var.outlook_redirect_uri
-  password_pepper                   = var.password_pepper
-  rabbit_mq_host                    = var.rabbit_mq_host
-  refresh_secret                    = var.refresh_secret
-  storage_container_sas_url         = var.storage_container_sas_url
-  tallulah_admin_password           = var.tallulah_admin_password
+  devops_keyvault_url               = data.azurerm_key_vault.keyvault_devops.vault_uri
+  jwt_secret                        = data.azurerm_key_vault_secret.keyvault_secrets["jwt-secret"].value
+  mongo_connection_url              = data.azurerm_key_vault_secret.keyvault_secrets["mongo-connection-url"].value
+  outlook_client_id                 = data.azurerm_key_vault_secret.keyvault_secrets["outlook-client-id"].value
+  outlook_client_secret             = data.azurerm_key_vault_secret.keyvault_secrets["outlook-client-secret"].value
+  outlook_tenant_id                 = data.azurerm_key_vault_secret.keyvault_secrets["outlook-tenant-id"].value
+  outlook_redirect_uri              = data.azurerm_key_vault_secret.keyvault_secrets["outlook-redirect-uri"].value
+  password_pepper                   = data.azurerm_key_vault_secret.keyvault_secrets["password-pepper"].value
+  rabbit_mq_host                    = data.azurerm_key_vault_secret.keyvault_secrets["rabbit-mq-host"].value
+  refresh_secret                    = data.azurerm_key_vault_secret.keyvault_secrets["refresh-secret"].value
+  storage_container_sas_url         = "TODO"
+  tallulah_admin_password           = data.azurerm_key_vault_secret.keyvault_secrets["tallulah-admin-password"].value
   storage_account_connection_string = module.storage_account.storage_account_connection_string
-  elastic_password                  = var.elastic_password
-  elastic_cloud_host                = var.elastic_cloud_host
-  openai_api_base                   = var.openai_api_base
-  openai_api_key                    = var.openai_api_key
-  email_no_reply_refresh_token      = var.email_no_reply_refresh_token
+  elastic_password                  = data.azurerm_key_vault_secret.keyvault_secrets["elastic-password"].value
+  elastic_cloud_host                = data.azurerm_key_vault_secret.keyvault_secrets["elastic-cloud-host"].value
+  openai_api_base                   = data.azurerm_key_vault_secret.keyvault_secrets["openai-api-base"].value
+  openai_api_key                    = data.azurerm_key_vault_secret.keyvault_secrets["openai-api-key"].value
+  email_no_reply_refresh_token      = data.azurerm_key_vault_secret.keyvault_secrets["email-no-reply-refresh-token"].value
 }
 
 module "container_app_rabbit_mq" {
@@ -112,10 +136,10 @@ module "container_app_rabbit_mq" {
   container_app_env_id        = module.container_apps_env.container_app_environment_id
   docker_image                = format("%s/%s", var.container_registry_server, var.rabbitmq_container_image_tag)
   container_registry_server   = var.container_registry_server
-  container_registry_username = var.container_registry_username
-  container_registry_password = var.container_registry_password
-  rabbit_mq_password          = var.rabbit_mq_password
-  rabbit_mq_user              = var.rabbit_mq_user
+  container_registry_username = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-username"].value
+  container_registry_password = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-password"].value
+  rabbit_mq_password          = data.azurerm_key_vault_secret.keyvault_secrets["rabbit-mq-password"].value
+  rabbit_mq_user              = data.azurerm_key_vault_secret.keyvault_secrets["rabbit-mq-user"].value
 }
 
 module "container_app_classifier" {
@@ -124,10 +148,10 @@ module "container_app_classifier" {
   container_app_env_id        = module.container_apps_env.container_app_environment_id
   docker_image                = format("%s/%s", var.container_registry_server, var.classifier_container_image_tag)
   container_registry_server   = var.container_registry_server
-  container_registry_username = var.container_registry_username
-  container_registry_password = var.container_registry_password
-  mongo_connection_url        = var.mongo_connection_url
-  rabbit_mq_host              = var.rabbit_mq_host
+  container_registry_username = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-username"].value
+  container_registry_password = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-password"].value
+  mongo_connection_url        = data.azurerm_key_vault_secret.keyvault_secrets["mongo-connection-url"].value
+  rabbit_mq_host              = data.azurerm_key_vault_secret.keyvault_secrets["rabbit-mq-host"].value
 }
 
 
@@ -137,8 +161,8 @@ module "container_app_frontend" {
   container_app_env_id        = module.container_apps_env.container_app_environment_id
   docker_image                = format("%s/%s", var.container_registry_server, var.ui_container_image_tag)
   container_registry_server   = var.container_registry_server
-  container_registry_username = var.container_registry_username
-  container_registry_password = var.container_registry_password
+  container_registry_username = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-username"].value
+  container_registry_password = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-password"].value
 }
 
 
@@ -148,11 +172,11 @@ module "container_app_logstash" {
   container_app_env_id        = module.container_apps_env.container_app_environment_id
   docker_image                = format("%s/%s", var.container_registry_server, var.logstash_container_image_tag)
   container_registry_server   = var.container_registry_server
-  container_registry_username = var.container_registry_username
-  container_registry_password = var.container_registry_password
-  elastic_cloud_username      = var.elastic_cloud_username
-  elastic_cloud_password      = var.elastic_cloud_password
-  elastic_cloud_host          = var.elastic_cloud_host
+  container_registry_username = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-username"].value
+  container_registry_password = data.azurerm_key_vault_secret.keyvault_secrets["container-registry-password"].value
+  elastic_cloud_username      = data.azurerm_key_vault_secret.keyvault_secrets["elastic-username"].value
+  elastic_cloud_password      = data.azurerm_key_vault_secret.keyvault_secrets["elastic-password"].value
+  elastic_cloud_host          = data.azurerm_key_vault_secret.keyvault_secrets["elastic-cloud-host"].value
 }