+++ devops/terraform/variables.tf
@@ -1,23 +1,3 @@
-variable "azure_tenant_id" {
-  description = "Azure Tenant ID which will have access to the Key Vault"
-  type        = string
-}
-
-variable "azure_object_id" {
-  description = "Azure Service Principal which will have access to the Key Vault"
-  type        = string
-}
-
-variable "azure_client_id" {
-  description = "Azure service principal client ID"
-  type        = string
-}
-
-variable "azure_client_secret" {
-  description = "Azure service principal client secret"
-  type        = string
-}
-
 variable "azure_subscription_id" {
   description = "Azure subscription ID"
   type        = string
@@ -28,116 +8,9 @@ variable "resource_group_name" {
   type        = string
 }
 
-variable "container_registry_password" {
-  description = "Container registry password"
-  type        = string
-  sensitive   = true
-}
-
 variable "container_registry_server" {
   description = "Container registry server"
   type        = string
-}
-
-variable "container_registry_username" {
-  description = "Container registry username"
-  type        = string
-}
-
-variable "jwt_secret" {
-  description = "JWT secret"
-  type        = string
-  sensitive   = true
-}
-
-variable "mongo_connection_url" {
-  description = "Mongodb connection url"
-  type        = string
-  sensitive   = true
-}
-
-variable "outlook_client_id" {
-  description = "Outlook client ID for mail integration"
-  type        = string
-}
-
-variable "outlook_client_secret" {
-  description = "Outlook client secret for mail integration"
-  type        = string
-  sensitive   = true
-}
-
-variable "outlook_redirect_uri" {
-  description = "Outlook redirect URI for mail integration"
-  type        = string
-}
-
-variable "outlook_tenant_id" {
-  description = "Outlook tenant ID for mail integration"
-  type        = string
-}
-
-variable "password_pepper" {
-  description = "Password pepper"
-  type        = string
-  sensitive   = true
-}
-
-variable "rabbit_mq_host" {
-  description = "RabbitMQ host"
-  type        = string
-}
-
-variable "rabbit_mq_password" {
-  description = "RabbitMQ password"
-  type        = string
-  sensitive   = true
-}
-
-variable "rabbit_mq_user" {
-  description = "RabbitMQ user"
-  type        = string
-}
-
-variable "refresh_secret" {
-  description = "Refresh secret"
-  type        = string
-  sensitive   = true
-}
-
-variable "storage_container_sas_url" {
-  description = "Storage container SAS URL"
-  type        = string
-}
-
-variable "tallulah_admin_password" {
-  description = "Tallulah admin password"
-  type        = string
-  sensitive   = true
-}
-
-variable "godaddy_secret" {
-  description = "value of godaddy secret"
-  type        = string
-  sensitive   = true
-}
-
-variable "godaddy_key" {
-  description = "value of godaddy key"
-  type        = string
-  sensitive   = true
-}
-
-variable "google_domains_token" {
-  description = "value of google domains token"
-  type        = string
-  sensitive   = true
-}
-
-variable "ssl_certificate_password" {
-  description = "SSL certificate password"
-  type        = string
-  sensitive   = true
 }
 
 variable "host_name" {
@@ -148,38 +21,6 @@ variable "host_name" {
 variable "host_name_2" {
   description = "Host name that the application gateway will listen to"
   type        = string
-}
-
-variable "elastic_cloud_username" {
-  description = "The username of the Elastic Cloud"
-  type        = string
-}
-
-variable "elastic_cloud_password" {
-  description = "The password of the Elastic Cloud"
-  type        = string
-  sensitive   = true
-}
-
-variable "elastic_cloud_host" {
-  description = "The URL of the Elastic Cloud"
-  type        = string
-}
-variable "elastic_password" {
-  description = "The password of the Elastic instance"
-  type        = string
-  sensitive   = true
-}
-
-variable "openai_api_base" {
-  description = "value of openai api base"
-  type        = string
-}
-
-variable "openai_api_key" {
-  description = "value of openai api key"
-  type        = string
-  sensitive   = true
 }
 
 variable "backend_container_image_tag" {
@@ -208,8 +49,17 @@ variable "classifier_container_image_tag" {
   type        = string
 }
 
-variable "email_no_reply_refresh_token" {
-  description = "The refresh token for the no-reply email account"
+variable "devops_keyvault_name" {
+  description = "The name of the keyvault"
   type        = string
-  sensitive   = true
+}
+
+variable "devops_keyvault_url" {
+  description = "The URL of the keyvault"
+  type        = string
+}
+
+variable "devops_resource_group_name" {
+  description = "The name of the resource group"
+  type        = string
 }