+++ scripts.sh
@@ -23,8 +23,6 @@ push_image_to_registry() {
         exit 1
     fi
 
-    echo "login to azure account"
-    az login --service-principal --username $AZURE_CLIENT_ID --password $AZURE_CLIENT_SECRET --tenant $AZURE_TENANT_ID
     az account set --subscription $AZURE_SUBSCRIPTION_ID
 
     echo "log in to azure registry"
@@ -132,6 +130,7 @@ generate_client() {
 }
 
 deploy() {
+    az login
     make build_image
     make push_all
 
@@ -155,19 +154,20 @@ deploy() {
     popd
 
     pushd devops/terraform
-    az account set --subscription $AZURE_SUBSCRIPTION_ID
-    az login
-    terraform workspace select development
     sed -i '' "s/^backend_container_image_tag=.*/backend_container_image_tag=\"tallulah\/backend:$backend_tag\"/g" development.tfvars
     sed -i '' "s/^ui_container_image_tag=.*/ui_container_image_tag=\"tallulah\/ui:$ui_tag\"/g" development.tfvars
     sed -i '' "s/^rabbitmq_container_image_tag=.*/rabbitmq_container_image_tag=\"tallulah\/rabbitmq:$backend_tag\"/g" development.tfvars
     sed -i '' "s/^logstash_container_image_tag=.*/logstash_container_image_tag=\"tallulah\/logstash:$backend_tag\"/g" development.tfvars
+
+    az account set --subscription $AZURE_SUBSCRIPTION_ID
+    terraform init -backend-config="backend.tfvars" -reconfigure
     terraform apply -var-file="development.tfvars" -auto-approve
     popd
 }
 
 
 release() {
+    az login
     make build_image
     make push_all
 
@@ -191,15 +191,16 @@ release() {
     popd
 
     pushd devops/terraform
-    az login --service-principal --username $AZURE_CLIENT_ID --password $AZURE_CLIENT_SECRET --tenant $AZURE_TENANT_ID
-    az account set --subscription $AZURE_SUBSCRIPTION_ID
-    az login
     sed -i '' "s/^backend_container_image_tag=.*/backend_container_image_tag=\"tallulah\/backend:$backend_tag\"/g" production.tfvars
     sed -i '' "s/^ui_container_image_tag=.*/ui_container_image_tag=\"tallulah\/ui:$ui_tag\"/g" production.tfvars
     sed -i '' "s/^rabbitmq_container_image_tag=.*/rabbitmq_container_image_tag=\"tallulah\/rabbitmq:$backend_tag\"/g" production.tfvars
     sed -i '' "s/^logstash_container_image_tag=.*/logstash_container_image_tag=\"tallulah\/logstash:$backend_tag\"/g" production.tfvars
-    # terraform workspace select default
-    # terraform apply -var-file="production.tfvars"
+
+    az account set --subscription $AZURE_SUBSCRIPTION_ID
+    terraform init -backend-config="backend_prod.tfvars" -reconfigure
+    # export AZURE_SUBSCRIPTION_ID="b7a46052-b7b1-433e-9147-56efbfe28ac5"
+    # az account set --subscription $AZURE_SUBSCRIPTION_ID
+    terraform apply -var-file="production.tfvars"
     popd
 }